requirements.txt
boto3==1.37.11
botocore==1.37.11
pandas==2.2.3
prettytable==3.16.0
tqdm==4.62.3
python-socketio[client]

# ===== AI Town Three-Layer Brain Dependencies =====

# IPC Communication
pyzmq>=25.0.0
tornado>=6.1  # Required for ZMQ on Windows

# LLM APIs
openai>=1.0.0
anthropic>=0.8.0
dashscope>=1.14.0  # Qwen API

# Vector Database for Experience Storage (optional for future RAG)
# chromadb>=0.4.22

# Text Embeddings (optional for future RAG)
# sentence-transformers>=2.2.2

# Data Validation
pydantic>=2.0.0

# Logging
loguru>=0.7.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0 